requirements.txt
attrs==22.1.0
Automat==20.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==37.0.4
cssselect==1.1.0
filelock==3.8.0
greenlet==1.1.2
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.7.0
itemloaders==1.0.4
jmespath==1.0.1
lxml==4.9.1
mysql==0.0.3
mysqlclient==2.1.1
numpy==1.23.2
pandas==1.4.3
parsel==1.6.0
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
PyMySQL==1.0.2
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pytz==2022.2.1
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
Scrapy==2.6.2
service-identity==21.1.0
six==1.16.0
SQLAlchemy==1.4.40
tldextract==3.3.1
Twisted==22.4.0
typing_extensions==4.3.0
urllib3==1.26.11
w3lib==2.0.1
zope.interface==5.4.0
attrs==22.1.0
Automat==20.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==37.0.4
cssselect==1.1.0
filelock==3.8.0
greenlet==1.1.2
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.7.0
itemloaders==1.0.4
jmespath==1.0.1
lxml==4.9.1
mysql==0.0.3
mysqlclient==2.1.1
numpy==1.23.2
pandas==1.4.3
parsel==1.6.0
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
PyMySQL==1.0.2
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
Scrapy==2.6.2
service-identity==21.1.0
six==1.16.0
SQLAlchemy==1.4.40
tldextract==3.3.1
Twisted==22.4.0
typing_extensions==4.3.0
urllib3==1.26.11
w3lib==2.0.1
zope.interface==5.4.0
